Add runtime manager to prevent long runtimes (#4)

+++ lib/RuntimeManager/RuntimeManager.cpp
@@ -22,7 +22,7 @@ void RuntimeManager::setMaxRuntime(unsigned int maxRuntime)
     _maxRuntime = maxRuntime;
 }
 
-void RuntimeManager::setMinWaitTime(unsigned long minWaitTime)
+void RuntimeManager::setMinWaitTime(unsigned int minWaitTime)
 {
     _minWaitTime = minWaitTime;
 }
